+++ crates/icy_board_engine/tests/predefined_functions.rs
@@ -66,6 +66,7 @@ async fn test_fmtreal(value: f64, field_width: i32, decimal_places: i32) -> Stri
         use_lmrs: true,
         cached_msg_header: None,
         abort_on_stack_error: true,
+        board_value: None,
         last_error: Default::default(),
         error_pending: false,
         error_handler: Default::default(),